+++ index-node.cjs.d.ts
@@ -1,2 +1,3 @@
 export * from "./dist/entry-node-cjs";
-export { default } from "./dist/entry-node-cjs";
+import aa from "./dist/entry-node-cjs";
+export default aa;