+++ lib/insights.ts
@@ -9,6 +9,7 @@ import { makeSendEvent, InsightsEventType, InsightsEvent } from "./_sendEvent";
 import { InitParams, init } from "./init";
 import { initSearch, InitSearchParams } from "./_initSearch";
 import { addAlgoliaAgent } from "./_algoliaAgent";
+import { getVersion } from "./_getVersion";
 
 import { RequestFnType } from "./utils/request";
 
@@ -87,6 +88,8 @@ class AlgoliaAnalytics {
   public init: (params: InitParams) => void;
   public initSearch: (params: InitSearchParams) => void;
 
+  public getVersion: (callback: (version: string) => void) => void;
+
   public addAlgoliaAgent: (algoliaAgent: string) => void;
 
   public setUserToken: (userToken: string) => void;
@@ -145,6 +148,8 @@ class AlgoliaAnalytics {
 
     this.viewedObjectIDs = viewedObjectIDs.bind(this);
     this.viewedFilters = viewedFilters.bind(this);
+
+    this.getVersion = getVersion.bind(this);
   }
 }